requirements.txt
bcrypt==4.3.0
fastapi==0.115.12
google-cloud-kms==3.5.1
google-cloud-logging==3.12.1
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
httpx==0.28.1
pydantic==2.11.5
pydantic-settings==2.9.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
uvicorn==0.34.3
